// src/tenable/types.ts
export interface RecentScanSummary {
  id: number;
  legacy: boolean;
  permissions: number;
  type: string;
  read: boolean;
  last_modification_date: number;
  creation_date: number;
  status: ScanStatus;
  uuid: string;
  shared: boolean;
  user_permissions: number;
  owner: string;
  schedule_uuid: string;
  timezone?: string;
  rrules?: string;
  starttime?: string;
  enabled: boolean;
  control: boolean;
  name: string;
}

export enum ScanStatus {
  Completed = 'completed',
  Aborted = 'aborted',
  Empty = 'empty',
  Imported = 'imported',
  Pending = 'pending',
  Running = 'running',
  Resuming = 'resuming',
  Canceling = 'canceling',
  Canceled = 'canceled',
  Pausing = 'pausing',
  Paused = 'paused',
  Stopping = 'stopping',
  Stopped = 'stopped',
}